refactor(agent-team): move the surface run into its module; drop a needless unsafe
Two gates caught real problems in the previous commit, which is the argument
for having them.

The file-length ratchet failed: inlining the surface phase into `main.rs`
pushed it to 582 lines. Moved into `surface::run`, where the rest of that
concern already lives, leaving `main.rs` at 498 and about orchestration. Also
corrected a stale doc comment there — it advertised "81 base tests, 94 with all
optional features" when the suite is 102.

Clippy and a grep caught the second: `rig-agent` had an `unsafe` block around
`std::env::set_var`. The workspace is edition 2021, where that call is safe, so
the block was pure noise — and this repository's only legitimate `unsafe` is
the counting allocator in `benches/`, per CONTRIBUTING. Removed.

Docs updated with the measured result: `examples/README.md` gains a surface
coverage table (all six examples at 44/44, with what a green run does *not*
mean spelled out), CHANGELOG and ROADMAP record all six rather than two, and
CONTRIBUTING's gate count goes 35 -> 38.

Verified: cargo run -p agent-team --release --all-features -> 102/102 tests,
44/44 cells, exit 0; cargo test --workspace --all-features clean; clippy
--all-features --all-targets -D warnings clean; fmt clean;
check_file_lengths exit 0.
